Split `common.rs` file into sub-files
