+++ src/unknown/network.rs
@@ -1,8 +1,6 @@
 // Take a look at the license at the top of the repository in the LICENSE file.
 
-use crate::common::MacAddr;
-use crate::IpNetwork;
-use crate::NetworkData;
+use crate::{IpNetwork, MacAddr, NetworkData};
 
 use std::collections::HashMap;
