+++ src/windows/network.rs
@@ -1,9 +1,7 @@
 // Take a look at the license at the top of the repository in the LICENSE file.
 
-use crate::common::MacAddr;
 use crate::network::refresh_networks_addresses;
-use crate::IpNetwork;
-use crate::NetworkData;
+use crate::{IpNetwork, MacAddr, NetworkData};
 
 use std::collections::{hash_map, HashMap};
