+++ src/windows/users.rs
@@ -1,11 +1,7 @@
 // Take a look at the license at the top of the repository in the LICENSE file.
 
 use crate::sys::utils::to_utf8_str;
-use crate::{
-    common::{Gid, Uid},
-    windows::sid::Sid,
-    Group, GroupInner, User,
-};
+use crate::{windows::sid::Sid, Gid, Group, GroupInner, Uid, User};
 
 use std::ptr::null_mut;
 use windows::core::{w, PCWSTR};
